+++ R/tokenizer-tokenizersbpe.R
@@ -10,7 +10,7 @@ tokenizers_bpe_tokens <- function(text, options = list()) {
   )
 
   if (length(options) > 0) {
-    token_expr <- mod_call_args(token_expr, args = options)
+    token_expr <- rlang::call_modify(token_expr, !!!options)
   }
 
   ddd <- utils::capture.output(type = "message", {