+++ NEWS.md
@@ -1,3 +1,5 @@
+# textrecipes (development version)
+
 # textrecipes 1.0.1
 
 * Examples no longer include `if (require(...))` code.